RAC-138: Split in smaller components, add file locale in product export controller
