+++ tests/front/unit/jest/unit.jest.js
@@ -9,6 +9,7 @@ const unitConfig = {
   coveragePathIgnorePatterns: [
     'src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/legacy-bridge',
     'src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/src/components/Button.tsx',
+    'src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/src/components/Modal.tsx',
     'src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/src/components/NoData.tsx',
     'src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/src/icons',
     'src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/src/illustrations',
